requirements.txt
selenium==4.18.1
undetected-chromedriver==3.5.5
pandas==2.2.1
python-dotenv==1.0.1
